+++ WpfApplication1.sln
@@ -3,12 +3,14 @@ Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio 14
 VisualStudioVersion = 14.0.22823.1
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "WpfApplication1", "WpfApplication1\WpfApplication1.csproj", "{C4F77D8E-EE5E-46CE-97DB-197E8C0BE0A8}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "AvaloniaSample", "AvaloniaSample\AvaloniaSample.csproj", "{C4F77D8E-EE5E-46CE-97DB-197E8C0BE0A8}"
 EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Solution Items", "Solution Items", "{D143FC46-7729-4AF3-8229-29704401EE73}"
 	ProjectSection(SolutionItems) = preProject
 		ReadMe.txt = ReadMe.txt
 	EndProjectSection
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Avalonia.Controls.VariableSizedWrapGrid", "Avalonia.Controls.VariableSizedWrapGrid\Avalonia.Controls.VariableSizedWrapGrid.csproj", "{D8535292-9BCC-4DFA-94BC-7B6483669081}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -38,6 +40,22 @@ Global
 		{C4F77D8E-EE5E-46CE-97DB-197E8C0BE0A8}.Release|x64.Build.0 = Release|Any CPU
 		{C4F77D8E-EE5E-46CE-97DB-197E8C0BE0A8}.Release|x86.ActiveCfg = Release|Any CPU
 		{C4F77D8E-EE5E-46CE-97DB-197E8C0BE0A8}.Release|x86.Build.0 = Release|Any CPU
+		{D8535292-9BCC-4DFA-94BC-7B6483669081}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{D8535292-9BCC-4DFA-94BC-7B6483669081}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{D8535292-9BCC-4DFA-94BC-7B6483669081}.Debug|ARM.ActiveCfg = Debug|Any CPU
+		{D8535292-9BCC-4DFA-94BC-7B6483669081}.Debug|ARM.Build.0 = Debug|Any CPU
+		{D8535292-9BCC-4DFA-94BC-7B6483669081}.Debug|x64.ActiveCfg = Debug|Any CPU
+		{D8535292-9BCC-4DFA-94BC-7B6483669081}.Debug|x64.Build.0 = Debug|Any CPU
+		{D8535292-9BCC-4DFA-94BC-7B6483669081}.Debug|x86.ActiveCfg = Debug|Any CPU
+		{D8535292-9BCC-4DFA-94BC-7B6483669081}.Debug|x86.Build.0 = Debug|Any CPU
+		{D8535292-9BCC-4DFA-94BC-7B6483669081}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{D8535292-9BCC-4DFA-94BC-7B6483669081}.Release|Any CPU.Build.0 = Release|Any CPU
+		{D8535292-9BCC-4DFA-94BC-7B6483669081}.Release|ARM.ActiveCfg = Release|Any CPU
+		{D8535292-9BCC-4DFA-94BC-7B6483669081}.Release|ARM.Build.0 = Release|Any CPU
+		{D8535292-9BCC-4DFA-94BC-7B6483669081}.Release|x64.ActiveCfg = Release|Any CPU
+		{D8535292-9BCC-4DFA-94BC-7B6483669081}.Release|x64.Build.0 = Release|Any CPU
+		{D8535292-9BCC-4DFA-94BC-7B6483669081}.Release|x86.ActiveCfg = Release|Any CPU
+		{D8535292-9BCC-4DFA-94BC-7B6483669081}.Release|x86.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
